Use ctx.view's EOF property for hollow groups
Essentially, the context's invoked_subcommand/subcommand_passed will
always be None here, because it is set in the superclass' invoke
coroutine. However, the hollow property really only mattered for groups
which had their group root command invoked, in the case of
invoke_without_command=False (which is the default). When this was true,
subcommand invocations would go directly to the subcommand, without any
checks on the root being executed. However, when this is false, the
group root "command" gets executed for checks and etc. however, at this
point, the command library is not aware of any subcommands yet (which is
what invoke() does at the beginning of its routine).

Here, we just check the StringView directly to see if we're at the end
of the message (which means no subcommand was passed).

Thank you @Ikusaba-san for the help.