Move command related classes and functions to lifesaver.commands
This is a huge breaking change.

Previously, the command and group decorators, alongside the Context
class and the Cog class were exposed in the lifesaver.bot module.
However, this doesn't make much sense.

They have been moved to the lifesaver.commands module (which is new) in
this patch. It would be inconvenient to import this module as it clashes
with discord.ext.commands (in code such as "from discord.ext import
commands", which is very common), so the children of lifesaver.commands
are also present in the lifesaver module.

The builtin extensions have been updated to account for this change. The
sample bot has also been improved when it came to testing some features.
MessageError and the module that it was contained in
(lifesaver.bot.errors) was also removed.