Manually implement lifesaver.utils.Ratelimiter
This is so that we don't have to rely on discord.py internals.

The semantics have also changed. You should use `Ratelimiter.hit` to
count requests and return whether the token is being rate limited. To
passively check if the token is being ratelimited without counting
towards the request count, use `Ratelimiter.is_being_rate_limited`.

Otherwise, the behavior should be identical.