Hot reload extension subfiles
Before, we used to use `os.path.listdir` to list the files (or folders)
in the extensions path and watch to detect changes. However, this didn't
really work for extensions that are modules (since there can be files in
subdirectories).

This patch makes `Poller` use `.glob("**/*")` to detect all files within
subdirectories. In order to correctly manipulate the correct extension,
the `PollerPlug` class now has some behavior to determine the extension
to manipulate based on the file modified. For example, if
`exts/mod/actions.py` is modified, `exts.mod` is reloaded -- not
`exts.mod.actions`.

The `Poller` setup behavior in `lifesaver.bot.bot.Bot` was also modified
to not fail silently (it was inside of a task).