Make progress updating types for Python 3.10
Bump the minimum version of the library to Python 3.10. There isn't a
good reason for this (I just wanted to use `type.__or__`), but it might
as well be communicated. A lot of the type hints have been updated for
Discord.py 2.0 (we were previously relying on the third-party type
stubs), but we still have a long way to go, especially if we want
everything to check out under pyright's strict mode.

I've figured out that Python type hinting is very clunky and flawed so
it's likely that these will never be perfect, but I want them to be good
enough eventually.